+++ src/topology/topology.hpp
@@ -5,7 +5,7 @@
 
 namespace cvm {
   namespace topology {
-    typedef uint64_t loc_t;
+    typedef uint32_t loc_t;
     extern loc_t null;
 
     std::vector<loc_t> get(const std::string& module);
